btrfs-progs: move print-tree.c to kernel-shared/
