+++ btrfs-find-root.c
@@ -27,7 +27,7 @@
 #include "kerncompat.h"
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
 #include "volumes.h"
