+++ btrfs-select-super.c
@@ -25,7 +25,7 @@
 #include "ctree.h"
 #include "volumes.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
 #include "common/utils.h"
