+++ check/qgroup-verify.c
@@ -25,7 +25,7 @@
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "common/utils.h"
 #include "kernel-shared/ulist.h"
 #include "common/rbtree-utils.h"
