+++ cmds/inspect-dump-super.c
@@ -24,7 +24,7 @@
 
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "common/utils.h"
 #include "cmds/commands.h"
 #include "common/help.h"
