+++ cmds/inspect-tree-stats.c
@@ -29,7 +29,7 @@
 #include "kerncompat.h"
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
 #include "volumes.h"
