+++ cmds/restore.c
@@ -39,7 +39,7 @@
 
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "kernel-lib/list.h"
 #include "volumes.h"
