+++ kernel-shared/extent-tree.c
@@ -25,7 +25,7 @@
 #include "kernel-lib/rbtree.h"
 #include "ctree.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "transaction.h"
 #include "crypto/crc32c.h"
 #include "volumes.h"
