+++ kernel-shared/uuid-tree.c
@@ -22,7 +22,7 @@
 #include "ctree.h"
 #include "transaction.h"
 #include "disk-io.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "common/utils.h"
 
 void btrfs_uuid_to_key(const u8 *uuid, struct btrfs_key *key)
