+++ volumes.c
@@ -25,7 +25,7 @@
 #include "ctree.h"
 #include "disk-io.h"
 #include "transaction.h"
-#include "print-tree.h"
+#include "kernel-shared/print-tree.h"
 #include "volumes.h"
 #include "common/utils.h"
 #include "kernel-lib/raid56.h"
